refactor healthcheck classes into separate folder. Update README.md and compose files
