+++ example/docker-compose.yml
@@ -3,7 +3,10 @@ services:
     container_name: renamarr
     image: ghcr.io/hollanbm/renamarr:latest@sha256:c11fae0fd5e808817c6f3639abce25a246ea560cf7099de9d8b498e61c0befa7
     # user: 1000:1000  # if using the log_to_file feature, recommended to set the container user to match the host user to avoid permission issues with log files.
+    read_only: true  # Restrict writes to explicit mounts.
     restart: unless-stopped
+    tmpfs:
+      - /tmp  # Required for the heartbeat with read_only enabled.
     volumes:
       - ./config.yml:/config/config.yml:ro
       # - ./logs:/logs:rw  # If using the log_to_file option, uncomment this line to persist logs to the host machine. **Don't forget to create logs folder first**
