+++ src/renamarr/healthcheck/health_state.py
@@ -1,0 +1,9 @@
+from enum import StrEnum
+
+
+class HealthState(StrEnum):
+    """Application lifecycle states exposed to the container health check."""
+
+    INITIALIZING = "initializing"
+    IDLE = "idle"
+    RUNNING = "running"
