+++ src/renamarr/healthcheck/settings.py
@@ -1,0 +1,7 @@
+from pathlib import Path
+
+
+HEALTH_FILE = Path("/tmp/renamarr-health.json")
+HEARTBEAT_INTERVAL_SECONDS = 10.0
+MAX_HEARTBEAT_AGE_SECONDS = 30.0
+SCHEMA_VERSION = 1
